Move GeoJSON to shared path
