+++ src/components/Guesser.tsx
@@ -9,7 +9,7 @@ import localeList from "../i18n/messages";
 import { FormattedMessage } from "react-intl";
 import { langNameMap } from "../i18n/locales";
 import { AltNames } from "../lib/alternateNames";
-const countryData: Country[] = require("../data/country_data.json").features;
+const countryData: Country[] = require("../../data/countries.geo.json").features;
 const alternateNames: AltNames = require("../data/alternate_names.json");
 
 type Props = {
